blur

+++ KnueppelKampfBase/Game/Components/MoveComponent.cs
@@ -7,13 +7,13 @@ namespace KnueppelKampfBase.Game.Components
 {
     public class MoveComponent : GameComponent
     {
-        const double MIN_VALUE = 0.001;
-        private double limit;
-        private double friction;
+        const float MIN_VALUE = 0.001;
+        private float limit;
+        private float friction;
         private Vector velocity;
         public Vector Velocity { get => velocity; set => velocity = value; }
-        public double Limit { get => limit; set => limit = value; }
-        public double Friction { get => friction; set => friction = value; }
+        public float Limit { get => limit; set => limit = value; }
+        public float Friction { get => friction; set => friction = value; }
 
         public MoveComponent()
         {
